requirements.txt
cryptography==1.7.1
idna==2.2
keyring==10.1
keyrings.alt==1.3
pyasn1==0.1.9
PyAudio==0.2.11
pycrypto==2.6.1
pygobject==3.22.0
python-apt==1.1.0b5
pyxdg==0.25
RPi.GPIO==0.6.5
SecretStorage==2.3.1
six==1.12.0
